dashboard/pyproject.toml
[tool.poetry]
name = "geodykes-dash"
version = "0.1.0"
description = "Dashboard to display dykes monitoring data"
authors = ["jurra <jose.urra86@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
dash = "^2.17.1"
dash-bootstrap-components = "^1.6.0"
pandas = "^2.2.2"
python-dotenv="*"
openapi_client = { path = "../sdk" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
